refactor modal func & add standup-request command
